refactor: 移动控制器到pipeline包
